* commit '9783b02747741f0d40845a87a6eb8d844cc73bff':
  fixes

+++ src/Brain/Cell/EntityResource/Common/Weight/WeightResourceInterface.php
@@ -17,7 +17,7 @@ interface WeightResourceInterface extends
     /**
      * Return the weight value.
      */
-    public function getValue(): int;
+    public function getValue(): float;
 
     /**
      * Return the weight unit.
